# api/protobuf/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/protobuf/api.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/call_back/transaction/notify": {
      "post": {
        "operationId": "NotifyService_TransactionNotify",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protobufTransactionNotifyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protobufTransactionNotifyRequest"
            }
          }
        ],
        "tags": [
          "NotifyService"
        ]
      }
    },
    "/v1/hello": {
      "get": {
        "operationId": "HelloService_GetHello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protobufHelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "tags": [
          "HelloService"
        ]
      }
    }
  },
  "definitions": {
    "HelloResponseData": {
      "type": "object",
      "properties": {
        "info": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "protobufHelloResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/HelloResponseData"
        }
      }
    },
    "protobufTransactionNotifyRequest": {
      "type": "object",
      "properties": {
        "requestType": {
          "type": "string",
          "format": "int64"
        },
        "requestId": {
          "type": "string"
        },
        "requestDetail": {
          "type": "string"
        },
        "extraInfo": {
          "type": "string"
        }
      }
    },
    "protobufTransactionNotifyResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "int64"
        },
        "requestId": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}